requirements.txt
telegram~=0.0.1
requests~=2.26.0
python-dotenv==0.19.0
python-telegram-bot==13.7
